requirements.txt
anthropic>=0.40.0
streamlit>=1.41.0
pymupdf>=1.25.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
